+++ src/main/scala/uk/gov/nationalarchives/tdr/api/service/DisplayPropertiesService.scala
@@ -32,7 +32,7 @@ class DisplayPropertiesService(displayPropertiesRepository: DisplayPropertiesRep
 
   def getDisplayProperties: Future[Seq[DisplayPropertyField]] = {
     for {
-      displayProperties <- displayPropertiesRepository.getDisplayProperties
+      displayProperties <- displayPropertiesRepository.getDisplayProperties()
     } yield {
       toDisplayProperty(displayProperties)
     }
@@ -40,7 +40,7 @@ class DisplayPropertiesService(displayPropertiesRepository: DisplayPropertiesRep
 
   def getActiveDisplayPropertyNames: Future[Seq[String]] = {
     for {
-      displayProperties <- displayPropertiesRepository.getDisplayProperties("Active", "true")
+      displayProperties <- displayPropertiesRepository.getDisplayProperties(Some("Active"), Some("true"))
     } yield {
       displayProperties.flatMap(_.propertyname)
     }
